requirements.txt
# LangChain ecosystem
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
langchain-ollama==0.3.6
langchain-text-splitters==0.3.9
rank_bm25>=0.2.2

# Vector DB
chromadb==1.0.15

# Embeddings
sentence-transformers==5.0.0
transformers==4.53.3
torch==2.7.1

# LLM runtime
ollama==0.5.1

# PDF support (Python 3.13 compatible)
pypdf==3.17.1
pdfplumber==0.10.3

# Utility
tqdm>=4.66.4
python-dotenv>=1.0.1
setuptools>=70.0.0
pip>=24.0
